+++ src/components/Animation/Type.tsx
