+++ src/hooks/useScrollTop.tsx
@@ -2,7 +2,7 @@
 import { useEffect } from "react";
 import { useLocation } from "react-router";
 
-const ScrollToTop = (props) => {
+const ScrollToTop = (props: any) => {
   const location = useLocation();
   useEffect(() => {
     window.scrollTo(0, 0);
