+++ src/pages/LandingPage.tsx
@@ -1,5 +1,5 @@
 import Hero from "../components/home/Hero";
-import Particles from "../components/Animation/Particles";
+// import Particles from "../components/Animation/Particles";
 import Footer from "../layouts/Footer";
 
 // import { useColorMode } from "@chakra-ui/react"
